refactor(friends): Add friends is now a fragment so it can be reused.
